Move config to init
